Move `Settings` and `ResolverSettings` to `ruff_workspace`
